# feat(P0 HOTFIX): Event-Pipeline production-ready – AFTER_COMMIT nur in Publishern, pg_notify abstrahiert, RBAC+Metriken, BATCH-Keys stabil