--- a/backend/backend/README.md
+++ b/backend/backend/README.md
@@ -1,0 +1,1 @@
+# Trigger CI