--- a/backend/backend/README.md
+++ b/backend/backend/README.md
@@ -1,1 +1,0 @@
-# Trigger CI