… README.md
… backend/README.md
… backend/docs/planung/adr/ADR-0002.md
… backend/pom.xml
… backend/src/main/java/de/freshplan/domain/cockpit/service/Metrics.java
… backend/src/main/java/de/freshplan/domain/cockpit/service/SalesCockpitService.java
… backend/src/main/java/de/freshplan/domain/cockpit/service/dto/LeadWidget.java
… backend/src/main/java/de/freshplan/domain/cockpit/service/dto/SalesCockpitDashboard.java
… backend/src/main/java/de/freshplan/domain/cockpit/service/query/SalesCockpitQueryService.java
… backend/src/main/java/de/freshplan/infrastructure/pg/HibernatePgNotifySender.java
… backend/src/main/java/de/freshplan/infrastructure/pg/PgNotifySender.java
… backend/src/main/java/de/freshplan/modules/leads/events/CrossModuleEventListener.java
… backend/src/main/java/de/freshplan/modules/leads/events/FollowUpProcessedEvent.java
… backend/src/main/java/de/freshplan/modules/leads/events/LeadEventPublisher.java
… backend/src/main/java/de/freshplan/modules/leads/events/LeadStatusChangeEvent.java
… backend/src/main/java/de/freshplan/modules/leads/service/FollowUpAutomationService.java
… backend/src/test/java/de/freshplan/domain/cockpit/DashboardAfterCommitTest.java
… backend/src/test/java/de/freshplan/domain/cockpit/DashboardBatchIdempotencyTest.java
… backend/src/test/java/de/freshplan/domain/cockpit/DashboardEventIdempotencyTest.java
… backend/src/test/java/de/freshplan/domain/cockpit/DashboardRBACTest.java
… backend/src/test/java/de/freshplan/domain/cockpit/DashboardTruncationTest.java
… backend/src/test/java/de/freshplan/infrastructure/pg/TestPgNotifySender.java
… backend/src/test/resources/application.properties
… docs/planung/CRM_COMPLETE_MASTER_PLAN_V5.md
… docs/planung/PRODUCTION_ROADMAP_2025.md
… docs/planung/SPRINT_2_1_1_DELTA_LOG.md
… docs/planung/TRIGGER_INDEX.md
… docs/planung/TRIGGER_SPRINT_2_1.md
… docs/planung/TRIGGER_SPRINT_2_1_1.md
… docs/planung/TRIGGER_SPRINT_2_2.md