## 🎯 Ziel

**P0 HOTFIX:** Integration Gaps zwischen Lead-Management und Dashboard beheben
- Event Distribution für Lead-Status-Changes und Follow-up Events
- Dashboard-Widget Integration mit Real-time Updates  
- Prometheus Metrics für Production Monitoring

## ⚠️ Risiko

**MEDIUM:** Event-System Änderungen könnten andere Module beeinflussen
- **Mitigation:** AFTER_COMMIT Pattern nur in Publishern, nicht in Listenern
- **Mitigation:** Idempotenz durch Caffeine Cache mit 500k Entries/24h TTL
- **Mitigation:** Fail-safe Error Handling - Fehler werden nicht propagiert

## 🔄 Migrations-Schritte + Rollback

**Keine Datenbank-Migration erforderlich!**
- Nur Code-Changes (Event Publisher/Listener)
- Rollback: Einfacher Revert des PRs möglich
- Event-System nutzt bestehende PostgreSQL LISTEN/NOTIFY

## ⚡ Performance-Nachweis

**P95 Latency: < 7ms** (validiert in LeadPerformanceValidationTest)
```
✅ testLeadCreationUnder10ms: P95 = 3.2ms (PASS)
✅ testStatusUpdateUnder10ms: P95 = 4.1ms (PASS) 
✅ testBulkOperationUnder200ms: 100 Leads in 142ms (PASS)
✅ testConcurrentAccessUnder50ms: P95 = 6.8ms (PASS)
✅ testSearchPerformanceUnder20ms: P95 = 5.3ms (PASS)
```

**Event Processing:** < 500ms Cross-Module Delivery
- NOTIFY Payload Guard: 8KB PostgreSQL Limit
- Dedupe Cache Hit-Rate: ~70% erwartet

## 🔒 Security-Checks

✅ **23 Security Tests** implementiert und bestanden:
- **RLS/ABAC:** @RlsContext auf allen Services
- **RBAC:** Role-based Checks in Publisher/Listener
- **Fail-Closed:** Explizite Deny bei fehlender Berechtigung
- **No Secrets:** Keine hardcoded Credentials
- **Input Validation:** Alle Events validiert

## 📚 SoT-Referenzen

- **Master Plan:** [CRM_COMPLETE_MASTER_PLAN_V5.md](docs/planung/CRM_COMPLETE_MASTER_PLAN_V5.md)
- **Trigger:** [TRIGGER_SPRINT_2_1_1.md](docs/planung/TRIGGER_SPRINT_2_1_1.md) 
- **Patterns:** 
  - [Security Test Pattern](docs/planung/features-neu/02_neukundengewinnung/artefakte/SECURITY_TEST_PATTERN.md)
  - [Performance Test Pattern](docs/planung/features-neu/02_neukundengewinnung/artefakte/PERFORMANCE_TEST_PATTERN.md)
  - [Event System Pattern](docs/planung/features-neu/02_neukundengewinnung/artefakte/EVENT_SYSTEM_PATTERN.md)

## ✅ Checklist

- [x] Code kompiliert ohne Fehler
- [x] Alle Tests laufen lokal durch (25 Event-Tests)
- [x] Spotless angewendet
- [x] Security-Checks durchgeführt
- [x] Performance-Benchmarks validiert
- [x] Code Review Findings adressiert
- [x] CI Pipeline läuft
- [ ] PR Review durch Team
- [ ] Merge nach main

## 🔄 Latest Updates

**2025-09-27:** Code Review Findings von Gemini adressiert
- HIGH Priority: Fail-fast mit Objects.requireNonNull(), Metrics-Optimierung
- MEDIUM Priority: UTC Timestamps, Cache-Konstanten, Record Cleanup
- CI Pipeline repariert (Workflow-Dateien aus main gemerged)