@joergstreeck 2025-09-26
## ✅ Final Documentation Sync Complete

**Docs synced across repo**: FP-235 & FP-236 überall als ✅ Complete markiert; PR #111 Inhalte (Envelope/UTC/Idempotenz/RBAC/Metriken) in Module 01-08 & `/infrastruktur` ergänzt; ADR-0002 und Delta-Logs aktualisiert; Konsistenz-Scans ohne kritische Alt-Artefakte.

### 📋 Key Updates:
- **ADR-0002** erweitert mit vollständigen Implementation Details:
  - PgNotifySender Abstraktion (Prod/Test)
  - AFTER_COMMIT Pattern (nur in Publishern)
  - 8KB Guard mit Truncation Handling
  - Caffeine Cache für Deduplizierung
  - Prometheus Metriken ohne _total suffix

- **PR Description** mit Production-Details aktualisiert
- **Alle Core-Dokumente** synchronisiert

### 🔍 Verifikation:
```bash
# Event-Flow testen
LISTEN dashboard_updates;

# Metriken prüfen
curl http://localhost:8080/q/metrics | grep freshplan_events

# PromQL für Grafana
sum by (result) (rate(freshplan_events_published{event_type="followup_processed"}[5m]))
```

### 📊 Production Ready:
- AFTER_COMMIT verhindert Ghost-Events
- Idempotenz via deterministischer Keys
- RBAC mit konfigurierbarem Test-Bypass
- Monitoring via Micrometer/Prometheus
- Performance P95 < 7ms validiert

**Reviewer-Hinweis**: Bitte einmal über die Modul-Dokus schauen, ob noch veraltete Snippets in Archiv-Ordnern sichtbar verlinkt sind (nicht blocking). Die 69 Fundstellen mit alten Patterns sind größtenteils in `/archiv` und nicht kritisch.

**Sprint 2.2 Kundenmanagement kann jetzt starten!** 🚀